feat: added get link and create order routes

+++ src/models/product.go
@@ -2,8 +2,8 @@ package models
 
 type Product struct {
 	Model
-	Title       string `json:"title"`
-	Description string `json:"description"`
-	Image       string `json:"image"`
-	Price       int    `json:"price"`
+	Title       string  `json:"title"`
+	Description string  `json:"description"`
+	Image       string  `json:"image"`
+	Price       float64 `json:"price"`
 }
